Wait for asterisk startup completion
Check that the service is ready to serve console connections
before considering it as "started".

This seems to be required by the FreePBX install script, as it seeks a
started and ready asterisk instance.

Systemd waits for asterisk up to 90s, then declares it "failed" and
attempts a new startup.